+++ tox.ini
@@ -40,11 +40,20 @@ deps =
   -r{toxinidir}/requirements/tests.txt
 whitelist_externals = find
                       docker
+                      sleep
 passenv = *_proxy
           *_PROXY
           DOCKER_*
 commands_pre =
-  docker run -d --name asyncpg_migrate_test -e POSTGRES_USER=root -e POSTGRES_PASSWORD=root -e POSTGRES_DB=asyncpg_migrate -p 5432:5432 postgres:alpine
+  docker run \
+    -d \
+    --name asyncpg_migrate_test \
+    -e POSTGRES_USER=root \
+    -e POSTGRES_PASSWORD=root \
+    -e POSTGRES_DB=asyncpg_migrate \
+    -p 5432:5432 \
+    postgres:alpine
+  sleep 5
 commands_post =
   docker stop asyncpg_migrate_test
   docker rm asyncpg_migrate_test
